+++ BCX_Wallet/library_base/src/main/java/com/cocos/library_base/entity/js_response/UpdateGasParamsModel.java
@@ -8,6 +8,6 @@ public class UpdateGasParamsModel {
 
     public String mortgager;
     public String beneficiary;
-    public int amount;
+    public double amount;
     public boolean isPropose;
 }
